modules/subscription_manager/requirements.txt
flask
paho-mqtt
redis
urllib3